Descriptive error message on HTTP Error 400.
When Linkedin returns an HTTP Error 400, the message that gets thrown with the LinkedInApiException is always `Client error response [url]** REQUEST_URL** [status code] 400 [reason phrase] Bad Request`.

A descriptive error message is encoded in the the XML body of the response, and can be found  in the `ClientException` sent by Guzzle. 

I added a new Catch block for the `ClientException` that parses the XML error body and throws it in the `LinkedInApiException` message.

Thanks a lot for your work on that library by the way! We're using it on a relatively large app and it's been a great help.