Suggestion
May I suggest the swapping of "app_id" for "client_id" and "app_secret" for "client_secret"? LinkedIn uses "client" instead of "app" in the "My Apps" section of the LinkedIn website. So I think that this change will make it more clear for anyone trying to use this API.